24.3.Sub Requests & Request Data {Adding the ?mac Override}
Adding the ?mac Override
Since I am using a Mac, it's kind of hard to test whether or not this feature correctly hides for people who are not on a Mac. To make testing easier, let's add a way to override the real logic. I want to be able to add a ?mac=true or ?mac=false to the URL to have full control.
The code for setting the argument is in IsMacArgumentValueResolver. So, if we want to "short-circuit" the real logic, it's no problem. Before we read the User-Agent, add if $request->query->has('mac'), then yield $request->query->getBoolean('mac). getBoolean() is a cool function that grabs the mac query parameter but runs it through PHP's filter_var() function with the FILTER_VALIDATE_BOOLEAN flag. That means a value like a false string will turn into a false boolean. Kinda fun. Anyways, after this, return so the function doesn't continue.
Ok: if I refresh without changing the URL, it still reads my User-Agent and everything looks right. Now add ?mac=false. And... it works! The message is gone. Oh wait! The first message is gone, but the one coming from the sub-request controller is still there! What the heck?
If you're thinking that somehow the argument value resolver isn't called on a sub request, that's not it. A sub request is handled exactly like the main request. This function is being called twice on this page: once for the main request and again for the sub request. So why do those two calls produce a different result?