24.4.Sub Requests & Request Data {Adding the ?mac Override}
The Request in the Sub Request is not the Same
Click into the profiler and go to the Performance section. The Request object that's being processed on top is not the same as the Request object that's being processed down here for the sub-request. Symfony creates two, distinct Request objects. The first Request object represents the data for the real HTTP request that's coming into our app. And so, it contains the query parameter info. But that second Request is kind of a "fake" request. It mainly exists so that the _controller attribute can be set on it. It's not really a representation of the "real" request. And so, it may not have all the same data. It doesn't have the query parameters, for example.
Let's see this: dump($request) inside of the resolve() method... then refresh
Hover over the target icon on the web debug toolbar. Yep, two dumps. If we look at the query parameters for the first Request... it's got it! mac=false. But down on the second request, it has some _path query parameter, but no mac.
The point is: there are two different requests. And the fact that they don't all contain the same data is on purpose. Because of this, whenever you're handling a sub-request, it's not a good idea to read information from the request... because you're not really reading data from the correct request!
So how can we correctly read the mac query parameter from a sub-request? To learn how, let's get crazy and make our own sub-request directly in PHP.