25.1.Manually Making a Sub Request
To understand more about sub requests, let's create one by hand! Because, it's not super obvious what these two Twig functions are really doing behind-the-scenes.
Insides our homepage controller, let's execute a sub request right here. How? It's simpler than you might think. Step 1: create a new request object: $request = new Request(). This is a totally empty Request object: it basically has nothing in it.
It's not like the Request::createFromGlobals() method that we saw earlier. That method pre-populates the object with all the current request information. This does not do that. To render the partial controller, set the request attribute: $request->attributes->set('_controller') and set that to the same string we have inside our Twig template. I'll copy that... and paste it here: 'App\Controller\PartialController::trendingQuotes'.
We now have a Request object with nothing in it except for an _controller attribute. And... that's all we need! Well, to work around some internal validation that checks for a valid IP address, we also need to say $request->server->set('REMOTE_ADDR', '127.0.0.1').
To send this into HttpKernel, we can fetch that service. Yes, even the mighty HttpKernel is a service in the container. Add another argument: HttpKernelInterface $httpKernel. Then, down here, we can say $response = ->handle(). We're going to pass this two arguments. We already know from index.php that the first argument is the Request. So, pass $request. But there is also an optional second argument: the request "type". This allows you to pass a flag that indicates if this is a "master" request - that's the default - or if this is a sub-request - some request that is happening inside the main one. That's our situation, so pass: HttpKernelInterface::SUB_REQUEST.
What difference will that make? Not much. But listeners to almost every event that we've seen are passed this flag on the event object and can behave differently based on whether or not a master or sub request is being handled. We'll see that in a few minutes.
To check if this works, dump($response).
Um... ok! Let's try this! We added this to the homepage... so refresh. Everything looks normal on this main request. Now hover over the target icon on the web debug toolbar. There it is! A dumped Response with the trending quotes content inside.
And, yes, if we click the time icon on the web debug toolbar to get to the Performance section of the profiler, we can see our sub request! Heck, now we have two sub requests: our "manual" sub-request and then the one from the template.
Set the threshold back down to 0 milliseconds. Way down on the main profiler, the sub-request shows up as this strange __section__.child thing.