25.2.Manually Making a Sub Request
Go back to the homepage controller and comment out the sub request logic.
I wanted you to see that this is all that really happens to trigger a sub request.

Listeners and the isMasterRequest() Flag
As we talked about, many listeners will use this SUB_REQUEST flag to change their behavior. Because sometimes, it only makes sense for a listener to do its work on the main, master request. For example - if you wrote a custom listener that checked the URL and denied access based on some custom logic, that listener only needs to do that check on the main request. It either denies access or allows access initially, and then the rest of the page should render normally.
Our UserAgentSubscriber is a perfect example of this. It makes no sense to read the User-Agent off of a sub request. It might work - because, in reality, sub-requests copy some of the data from the main request, but trying to read real information off of the request in a sub-request is asking for trouble. I really want you to think of the master and sub requests as totally independent objects.
So, what can we do? At the very top of our listener, if not $event->isMasterRequest(), simply return.
The isMasterRequest() method is a shortcut to check the flag that was originally passed to HttpKernel::handle(). Our listener will still be called on a sub-request, but now it will do nothing. And that makes sense: this class is doing nothing more than logging the User-Agent. We didn't realize it before, but thanks to our sub-request, each page refresh was logging the User-Agent twice: one for the main request and once for the sub-request.
Ok, but! We still haven't fixed our original problem: when we add ?mac=false to the URL, this is correctly read on the master request but incorrectly on the sub request. That's because we're trying to read that query parameter from inside the sub request... which doesn't work.
How can we fix that? The answer leverages an old friend of ours and will also touch on the proper way to pass data from the main request to the sub-request if you want to use HTTP caching with edge side includes. That's next.