26.1.Sub Request Attributes {Passing isMac through the Request Attributes}
We have this problem: we know the cause, but not the solution. Basically, we're reading this ?isMac=false from both our master request - where it's being read correctly - and our sub-request - where it is not. That's because those are two separate request objects. And, in general, we should avoid reading request info from a sub-request... because we're not working with the real Request object!
But accidentally trying to use data from the real request in a sub-request was so easy to do! Inside of PartialController, we simply added an $isMac argument like we do everywhere else. This argument was passed thanks to our custom IsMacArgumentValueResolver: this is executed on every request. But the second time it's called, the request is not the real request.
Here's the plan: I want to prevent any controllers that are being called on a sub-request from being able to have an $isMac argument. If we try, I want an exception to be thrown.
By the way, it is actually possible to use the RequestStack from anywhere to get the "master" request: it has a getMasterRequest() method. But! If you're using HTTP caching with edge side includes, this will break. The solution we're going to show is the proper one... and a lot more fun to implement for a deep dive.

Passing isMac through the Request Attributes
The easiest way to do this would be to go to our argument value resolver, find out if this is a sub-request and return false from supports() in that case. But... there's not a great way to figure out if this is the master or sub-request from here. That's mostly a superpower of event listeners.
So, let's use a trick: let's allow the listener to figure out whether or not we're using a Mac. And then, it can pass that information over to our argument value resolver.
Inside the subscriber, create a new private function called isMac(): it will take in the Request object from HttpFoundation and return a boolean. For the logic, copy everything from inside of resolve(), remove it and paste it here. Change both of the yield calls to return.