26.2.Sub Request Attributes {Passing isMac through the Request Attributes}
Perfect! We now have a function to tell us if we're using a Mac. Now, up in onKernelRequest(), if this is a sub-request, the method returns immediately. But if it's a master request, let's work some magic. How can we pass the isMac value to the argument value resolver? We could create and call a setter on it. Or, we could put that into the request attributes! I mean, that is the proper place for information about the request that is specific to your app. This is a perfect example!
Do it with $request->attributes->set() and create a key called _isMac set to $this->isMac($request).
We're calling it _isMac because if we called it just isMac, we wouldn't even need the argument value resolver! It would immediately be possible to have an $isMac argument without an error. So... I want to try to do this, kind of, the hard way.
Move over to the IsMacArgumentValueResolver. Here, we're going to read that attribute, which will only exist on the master request. Inside of supports(), add && $request->attributes->has('_isMac'). Supports will now return false for a sub request.
In resolve, yield $request->attributes->get('_isMac')
That's it! Does it work? In your browser, first open an article show page in a new tab. Here, it works: you can see it on the web debug toolbar: we're dumping the $isMac argument in the controller. This proves that we haven't broken the master request.

But now, refresh the homepage. It explodes!

  An exception has been throwing during the rendering of a template: could not resolve $isMac argument.

And you can see that it hits this on our {{ render(controller()) }} line. This is exactly what I wanted: I can no longer accidentally use this argument: it only works on the master request.