26.3.Sub Request Attributes {Passing Attributes Directly to a Sub-Request}
Passing Attributes Directly to a Sub-Request
Of course... the question now is: how do we fix this? What if I really do need to know the $isMac value from inside a sub request? I mean, that's a valid thing to want to know!
The solution is to pass the isMac value from your main request to your sub request object. In a Twig template, you can do this by passing an optional second array argument. Pass isMac set to isMac.
Before we chat about this, let's see if it works. Refresh and... it does! Woh! Let... me explain. The second argument to controller() is an array of values that you want to pass to the attributes of the sub-request object. Why does that work? Well, for the sub-request, the $isMac controller argument is actually not being passed to us thanks to the IsMacArgumentValueResolver. Nope! It works simply because we put an isMac key into the request attributes... and anything inside request attributes are allowed as controller arguments.
It feels like the second argument to the Twig controller() function is just an "array of variables to pass to the controller". And while that's essentially true, it really works thanks to the request attributes.
Back on the homepage, because we still have the ?mac=false query parameter, the message is not rendered in both places. If we remove that query parameter... yes! It's displayed consistently.