26.4.Sub Request Attributes {Attributes and Edge Side Includes}
Attributes and Edge Side Includes
As an added benefit, this was of passing variables from the master to the sub-request works perfectly with edge side includes and HTTP caching. Open the profiler, go to the Performance tab, find the sub request and click to go into its profiler. See this internal URL up here? If we were truly using edge side includes, this URL would be a real URL used by your HTTP cache to fetch this fragment and the URL would have an extra isMac=true or false query parameter. For ESI, request attributes becomes part of the URL. And since the URL operates as the cache key for HTTP caching, this means that you would have a separate cache for the version of your trending quotes with the "isMac" message and without. That's perfect.
Close up that tab. Now that we know that this array become request attributes, another solution would be to pass _isMac. If we did that, it would be used by our argument value resolver. UserAgentSubscriber would still do nothing because it's a sub request... but the _isMac attribute would be there and the IsMacArgumentValueResolver would make the isMac argument available to PartialController. Two different ways to do the exact same thing... and both are super nerdy.
Ok friends! Wow! That was fun! I hope you enjoyed this deep dive into the heart of Symfony's HttpKernel as much as I did. If you have still have questions or want to deep-dive into a different part of Symfony, let us know in the comments.
Happy hacking! And we'll see ya next time.